Add authorization

+++ api/users.go
@@ -10,7 +10,7 @@ import (
 func initUserHandlers(userRouter fiber.Router) {
 	userRouter.Post("/", userCreateHandler())
 
-	userRouter.Use(auth.CheckToken())
+	userRouter.Use(auth.CheckToken(db.UserUT))
 	userRouter.Get("/", userListHandler())
 	userRouter.Get("/:id", userDetailHandler())
 	userRouter.Put("/:id", userUpdateHandler())
@@ -78,6 +78,7 @@ func userCreateHandler() func(*fiber.Ctx) error {
 		if err := c.BodyParser(u); err != nil {
 			return fiber.NewError(fiber.StatusBadRequest, err.Error())
 		}
+		u.Type = db.UserUT
 
 		hashedPW, err := auth.HashPW(u.Password)
 		if err != nil {
@@ -89,6 +90,7 @@ func userCreateHandler() func(*fiber.Ctx) error {
 		if err != nil {
 			return fiber.NewError(fiber.StatusInternalServerError, err.Error())
 		}
+		user.Password = ""
 
 		c.JSON(user)
 		return c.SendStatus(201)
@@ -118,12 +120,38 @@ func userUpdateHandler() func(*fiber.Ctx) error {
 				"Could not convert given object ID, did you use the right ID-format?")
 		}
 
+		// only super-admin or the user themself are allowed to edit
+		isSuperAdmin := c.Locals("jwtUserType") == db.SuperAdminUT
+		if id != c.Locals("jwtUserId") && !isSuperAdmin {
+			return fiber.NewError(fiber.StatusUnauthorized)
+		}
+
+		// one may not change other users passwords
+		if id != c.Locals("jwtUserId") {
+			u.Password = ""
+		}
+
+		// only super admins may change user types
+		if !isSuperAdmin {
+			u.Type = c.Locals("jwtUserType").(db.UserType)
+		}
+
+		// if a user wants to change their PW, hash it first
+		if u.Password != "" {
+			u.Password, err = auth.HashPW(u.Password)
+			if err != nil {
+				return fiber.NewError(fiber.StatusBadRequest,
+					"Could not hash the given password : "+err.Error())
+			}
+		}
+
 		user, err := db.EditUser(id, *u)
+
 		if err != nil {
 			return fiber.NewError(fiber.StatusNotFound, err.Error())
 		}
 
-		user.ID = id
+		user.Password = ""
 		c.JSON(user)
 		return c.SendStatus(200)
 	}
@@ -143,6 +171,10 @@ func userDeleteHandler() func(*fiber.Ctx) error {
 		if err != nil {
 			return fiber.NewError(fiber.StatusBadRequest,
 				"Could not convert given object ID, did you use the right ID-format?")
+		}
+
+		if id != c.Locals("jwtUserId") {
+			return fiber.NewError(fiber.StatusUnauthorized)
 		}
 
 		err = db.DeleteUser(id)
